include/arch: make uintptr/intptr pointer-wide on amd64 and arm64
Both had `typedef unsigned long uintptr`, which reads as obviously right
and is wrong on exactly those two arches: every Plan 9 C compiler in this
tree defines `long` as 4 bytes (SZ_LONG in compilers/*c/gc.h) regardless
of arch, including 6c and 7c, while pointers there are 8 (SZ_IND). So
uintptr was 4 bytes on a 64-bit arch and silently dropped the top 32 bits
of any pointer cast through it.

Confirmed by probe rather than by reading gc.h: `sizeof(uintptr)` through
6c -S / 7c -S emitted $4 next to `sizeof(void*)`'s $8.

riscv64/u.h had already hit this and used `unsigned long long` -- but its
comment asserted the other two were fine "because those two *do* have an
8-byte long", which is what kept anyone from checking them. That claim is
corrected here too; it would otherwise talk the next reader out of
looking. Same for the adjacent FPdbleword comments on amd64/arm64, which
justified using u32 for hi/lo by claiming "ulong is 8 bytes on this arch"
-- it is 4, so ulong would happen to work; u32 is still the right way to
write it, just not for that reason.

Nothing caught this because the two places that actually round-trip a
pointer through uintptr (port/getcallerpc.c's return value, fmt/fmtfd.c's
(void*)(uintptr)fd) both survive truncation in a statically linked binary,
whose text and data addresses all fit in 32 bits anyway. Found while
starting port/sbrk.c, whose break arithmetic goes through uintptr.

tests/c/regressions/uintptr_width.c gates it, on arm64/amd64/riscv64 --
riscv64 was already correct and is the control, which is the point of
testing an invariant rather than a fix. Verified to actually catch the
bug: putting the old typedef back makes amd64 exit 1 while the other two
still pass. Worth knowing 6c also reports it directly ("warning:
conversion of pointer to shorter integer"), so it was diagnosable all
along -- it predates warnings being visible by default.

Full mk test passes.

+++ include/arch/arm64/u.h
@@ -13,15 +13,21 @@ typedef unsigned long long u64;
 typedef float float32;
 typedef double float64;
 
-//TODO? not unsigned long long like in riscv64/u.h? see the comment there
-typedef unsigned long uintptr;
-typedef long intptr;
+// claude: deliberately vlong-width, like riscv64/u.h's and amd64/u.h's
+// (see the latter for the full reasoning). 7c's `long` is 4 bytes
+// (SZ_LONG in compilers/7c/gc.h) while pointers are 8 (SZ_IND), so
+// `unsigned long` here would make uintptr 4 bytes on a 64-bit arch and
+// truncate every pointer cast through it.
+typedef unsigned long long uintptr;
+typedef long long intptr;
 
 // bit-level double access for port/frexp.c (frexp/ldexp/modf), adapted
 // from principia's include/arch/{arm,386}/u.h (same little-endian
-// layout) -- note this deliberately uses uint32 for hi/lo, not ulong:
-// ulong is 8 bytes on this arch (unlike 32-bit arm/386, where it's the
-// same size as uint32), which would silently double the union's size.
+// layout) -- note this deliberately uses u32 for hi/lo, not ulong: u32
+// states the exact width this bit layout requires, rather than relying
+// on whatever `ulong` happens to be. claude: `ulong` is in fact also 4
+// bytes here (7c's `long` is 4 -- see the uintptr note above). See
+// amd64/u.h's identical note.
 union FPdbleword {
 	double x;
 	struct {	/* little endian */
